Imported Upstream version 0.4.0

+++ wscript
@@ -1,7 +1,8 @@
 #!/usr/bin/env python
 import re
 import Options
-import sys, os, shutil
+import sys, os, shutil, glob
+import Utils
 from Utils import cmd_output
 from os.path import join, dirname, abspath
 from logging import fatal
@@ -9,16 +10,29 @@ from logging import fatal
 cwd = os.getcwd()
 APPNAME="node.js"
 
+# Use the directory that this file is found in to find the tools
+# directory where the js2c.py file can be found.
+sys.path.append(sys.argv[0] + '/tools');
 import js2c
 
 srcdir = '.'
 blddir = 'build'
-
+supported_archs = ('arm', 'ia32', 'x64') # 'mips' supported by v8, but not node
 
 jobs=1
 if os.environ.has_key('JOBS'):
   jobs = int(os.environ['JOBS'])
 
+def safe_path(path):
+  return path.replace("\\", "/")
+
+def canonical_cpu_type(arch):
+  m = {'x86': 'ia32', 'i386':'ia32', 'x86_64':'x64', 'amd64':'x64'}
+  if arch in m: arch = m[arch]
+  if not arch in supported_archs:
+    raise Exception("supported architectures are "+', '.join(supported_archs)+\
+                    " but NOT '" + arch + "'.")
+  return arch
 
 def set_options(opt):
   # the gcc module provides a --debug-level option
@@ -87,6 +101,20 @@ def set_options(opt):
                 , dest='shared_v8_libname'
                 )
 
+  opt.add_option( '--oprofile'
+                , action='store_true'
+                , default=False
+                , help="add oprofile support"
+                , dest='use_oprofile'
+                )
+
+  opt.add_option( '--gdb'
+                , action='store_true'
+                , default=False
+                , help="add gdb support"
+                , dest='use_gdbjit'
+                )
+
 
   opt.add_option('--shared-cares'
                 , action='store_true'
@@ -131,6 +159,30 @@ def set_options(opt):
                 , dest='shared_libev_libpath'
                 )
 
+  opt.add_option( '--with-dtrace'
+                , action='store_true'
+                , default=False
+                , help='Build with DTrace (experimental)'
+                , dest='dtrace'
+                )
+ 
+
+  opt.add_option( '--product-type'
+                , action='store'
+                , default='program'
+                , help='What kind of product to produce (program, cstaticlib '\
+                       'or cshlib) [default: %default]'
+                , dest='product_type'
+                )
+
+  opt.add_option( '--dest-cpu'
+                , action='store'
+                , default=None
+                , help='CPU architecture to build for. Valid values are: '+\
+                       ', '.join(supported_archs)
+                , dest='dest_cpu'
+                )
+
 
 
 
@@ -143,23 +195,40 @@ def configure(conf):
   o = Options.options
 
   conf.env["USE_DEBUG"] = o.debug
-  conf.env["SNAPSHOT_V8"] = not o.without_snapshot
+  # Snapshot building does noet seem to work on cygwin and mingw32
+  conf.env["SNAPSHOT_V8"] = not o.without_snapshot and not sys.platform.startswith("cygwin") and not sys.platform.startswith("win32")
+  if sys.platform.startswith("sunos"):
+    conf.env["SNAPSHOT_V8"] = False
   conf.env["USE_PROFILING"] = o.profile
 
   conf.env["USE_SHARED_V8"] = o.shared_v8 or o.shared_v8_includes or o.shared_v8_libpath or o.shared_v8_libname
   conf.env["USE_SHARED_CARES"] = o.shared_cares or o.shared_cares_includes or o.shared_cares_libpath
   conf.env["USE_SHARED_LIBEV"] = o.shared_libev or o.shared_libev_includes or o.shared_libev_libpath
 
+  conf.env["USE_OPROFILE"] = o.use_oprofile
+  conf.env["USE_GDBJIT"] = o.use_gdbjit
+
+  if o.use_oprofile:
+    conf.check(lib=['bfd', 'opagent'], uselib_store="OPROFILE")
+
   conf.check(lib='dl', uselib_store='DL')
-  if not sys.platform.startswith("sunos") and not sys.platform.startswith("cygwin"):
+  if not sys.platform.startswith("sunos") and not sys.platform.startswith("cygwin") and not sys.platform.startswith("win32"):
     conf.env.append_value("CCFLAGS", "-rdynamic")
     conf.env.append_value("LINKFLAGS_DL", "-rdynamic")
 
-  if sys.platform.startswith("freebsd"):
+  if sys.platform.startswith("freebsd") or sys.platform.startswith("openbsd"):
     conf.check(lib='kvm', uselib_store='KVM')
 
   #if Options.options.debug:
   #  conf.check(lib='profiler', uselib_store='PROFILER')
+
+  if Options.options.dtrace:
+    if not sys.platform.startswith("sunos"):
+      conf.fatal('DTrace support only currently available on Solaris')
+
+    conf.find_program('dtrace', var='DTRACE', mandatory=True)
+    conf.env["USE_DTRACE"] = True
+    conf.env.append_value("CXXFLAGS", "-DHAVE_DTRACE=1")
 
   if Options.options.efence:
     conf.check(lib='efence', libpath=['/usr/lib', '/usr/local/lib'], uselib_store='EFENCE')
@@ -196,6 +265,17 @@ def configure(conf):
   else:
     Options.options.use_openssl = conf.env["USE_OPENSSL"] = False
 
+  conf.check(lib='util', libpath=['/usr/lib', '/usr/local/lib'],
+             uselib_store='UTIL')
+
+  # normalize DEST_CPU from --dest-cpu, DEST_CPU or built-in value
+  if Options.options.dest_cpu and Options.options.dest_cpu:
+    conf.env['DEST_CPU'] = canonical_cpu_type(Options.options.dest_cpu)
+  elif 'DEST_CPU' in os.environ and os.environ['DEST_CPU']:
+    conf.env['DEST_CPU'] = canonical_cpu_type(os.environ['DEST_CPU'])
+  elif 'DEST_CPU' in conf.env and conf.env['DEST_CPU']:
+    conf.env['DEST_CPU'] = canonical_cpu_type(conf.env['DEST_CPU'])
+
   conf.check(lib='rt', uselib_store='RT')
 
   if sys.platform.startswith("sunos"):
@@ -228,7 +308,12 @@ def configure(conf):
                             libpath=v8_libpath):
         conf.fatal("Cannot find v8_g")
 
-  if conf.env['USE_SHARED_CARES']:
+  if sys.platform.startswith("win32"):
+    # On win32 CARES is always static, so we can call internal functions like ares_inet_pton et al. 
+    # CARES_STATICLIB must be defined or gcc will try to make DLL stub calls
+    conf.env.append_value('CPPFLAGS', '-DCARES_STATICLIB=1')
+    conf.sub_config('deps/c-ares')
+  elif conf.env['USE_SHARED_CARES']:
     cares_includes = [];
     if o.shared_cares_includes: cares_includes.append(o.shared_cares_includes);
     cares_libpath = [];
@@ -264,7 +349,7 @@ def configure(conf):
     conf.env.append_value ('CCFLAGS', '-threads')
     conf.env.append_value ('CXXFLAGS', '-threads')
     #conf.env.append_value ('LINKFLAGS', ' -threads')
-  elif not sys.platform.startswith("cygwin"):
+  elif not sys.platform.startswith("cygwin") and not sys.platform.startswith("win32"):
     threadflags='-pthread'
     conf.env.append_value ('CCFLAGS', threadflags)
     conf.env.append_value ('CXXFLAGS', threadflags)
@@ -272,12 +357,38 @@ def configure(conf):
   if sys.platform.startswith("darwin"):
     # used by platform_darwin_*.cc
     conf.env.append_value('LINKFLAGS', ['-framework','Carbon'])
+    # cross compile for architecture specified by DEST_CPU
+    if 'DEST_CPU' in conf.env:
+      arch = conf.env['DEST_CPU']
+      # map supported_archs to GCC names:
+      arch_mappings = {'ia32': 'i386', 'x64': 'x86_64'}
+      if arch in arch_mappings:
+        arch = arch_mappings[arch]
+      flags = ['-arch', arch]
+      conf.env.append_value('CCFLAGS', flags)
+      conf.env.append_value('CXXFLAGS', flags)
+      conf.env.append_value('LINKFLAGS', flags)
+  if 'DEST_CPU' in conf.env:
+    arch = conf.env['DEST_CPU']
+    # TODO: -m32 is only available on 64 bit machines, so check host type
+    flags = None
+    if arch == 'ia32':
+      flags = '-m32'
+    if flags:
+      conf.env.append_value('CCFLAGS', flags)
+      conf.env.append_value('CXXFLAGS', flags)
+      conf.env.append_value('LINKFLAGS', flags)
 
   # Needed for getaddrinfo in libeio
   conf.env.append_value("CPPFLAGS", "-DX_STACKSIZE=%d" % (1024*64))
   # LFS
   conf.env.append_value('CPPFLAGS',  '-D_LARGEFILE_SOURCE')
   conf.env.append_value('CPPFLAGS',  '-D_FILE_OFFSET_BITS=64')
+  conf.env.append_value('CPPFLAGS',  '-DEV_MULTIPLICITY=0')
+
+  # Makes select on windows support more than 64 FDs
+  if sys.platform.startswith("win32"):
+    conf.env.append_value('CPPFLAGS', '-DFD_SETSIZE=1024');
 
   ## needed for node_file.cc fdatasync
   ## Strangely on OSX 10.6 the g++ doesn't see fdatasync but gcc does?
@@ -298,13 +409,39 @@ def configure(conf):
   # platform
   conf.env.append_value('CPPFLAGS', '-DPLATFORM="' + conf.env['DEST_OS'] + '"')
 
+  # posix?
+  if not sys.platform.startswith('win'):
+    conf.env.append_value('CPPFLAGS', '-D__POSIX__=1')
+
+  platform_file = "src/platform_%s.cc" % conf.env['DEST_OS']
+  if os.path.exists(join(cwd, platform_file)):
+    Options.options.platform_file = True
+    conf.env["PLATFORM_FILE"] = platform_file
+  else:
+    Options.options.platform_file = False
+    conf.env["PLATFORM_FILE"] = "src/platform_none.cc"
+
   if conf.env['USE_PROFILING'] == True:
     conf.env.append_value('CPPFLAGS', '-pg')
     conf.env.append_value('LINKFLAGS', '-pg')
 
+  if sys.platform.startswith("win32"):
+    conf.env.append_value('LIB', 'ws2_32')
+    conf.env.append_value('LIB', 'winmm')
+
+  conf.env.append_value('CPPFLAGS', '-Wno-unused-parameter');
+  conf.env.append_value('CPPFLAGS', '-D_FORTIFY_SOURCE=2');
+
   # Split off debug variant before adding variant specific defines
   debug_env = conf.env.copy()
   conf.set_env_name('debug', debug_env)
+
+  if (sys.platform.startswith("win32")):
+    # Static pthread - crashes
+    #conf.env.append_value('LINKFLAGS', '../deps/pthreads-w32/libpthreadGC2.a')
+    #debug_env.append_value('LINKFLAGS', '../deps/pthreads-w32/libpthreadGC2d.a')
+    # Pthread dll
+    conf.env.append_value('LIB', 'pthread.dll')
 
   # Configure debug variant
   conf.setenv('debug')
@@ -334,10 +471,11 @@ def v8_cmd(bld, variant):
   # cannot see symbols in the executable which are hidden, even if the
   # executable is statically linked together...
 
-  # XXX Remove this when v8 defaults x86_64 to native builds
+  # XXX Change this when v8 defaults x86_64 to native builds
+  # Possible values are (arm, ia32, x64, mips).
   arch = ""
-  if bld.env['DEST_CPU'] == 'x86_64':
-    arch = "arch=x64"
+  if bld.env['DEST_CPU']:
+    arch = "arch="+bld.env['DEST_CPU']
 
   if variant == "default":
     mode = "release"
@@ -349,16 +487,25 @@ def v8_cmd(bld, variant):
   else:
     snapshot = ""
 
-  cmd_R = 'python "%s" -j %d -C "%s" -Y "%s" visibility=default mode=%s %s library=static %s'
+  if bld.env["USE_OPROFILE"]:
+    profile = "prof=oprofile"
+  else:
+    profile = ""
+
+  cmd_R = sys.executable + ' "%s" -j %d -C "%s" -Y "%s" visibility=default mode=%s %s library=static %s %s'
 
   cmd = cmd_R % ( scons
                 , Options.options.jobs
-                , bld.srcnode.abspath(bld.env_of_name(variant))
-                , v8dir_src
+                , safe_path(bld.srcnode.abspath(bld.env_of_name(variant)))
+                , safe_path(v8dir_src)
                 , mode
                 , arch
                 , snapshot
+                , profile
                 )
+
+  if bld.env["USE_GDBJIT"]:
+    cmd += ' gdbjit=on '
 
   if sys.platform.startswith("sunos"): cmd += ' toolchain=gcc'
 
@@ -404,10 +551,13 @@ def build(bld):
   Build.BuildContext.exec_command = exec_command
 
   Options.options.jobs=jobs
+  product_type = Options.options.product_type
+  product_type_is_lib = product_type != 'program'
 
   print "DEST_OS: " + bld.env['DEST_OS']
   print "DEST_CPU: " + bld.env['DEST_CPU']
   print "Parallel Jobs: " + str(Options.options.jobs)
+  print "Product type: " + product_type
 
   bld.add_subdirs('deps/libeio')
 
@@ -428,7 +578,7 @@ def build(bld):
 
   ### src/native.cc
   def make_macros(loc, content):
-    f = open(loc, 'w')
+    f = open(loc, 'a')
     f.write(content)
     f.close
 
@@ -442,9 +592,32 @@ def build(bld):
     "macros.py"
   )
 
+  ### We need to truncate the macros.py file
+  f = open(macros_loc_debug, 'w')
+  f.close
+  f = open(macros_loc_default, 'w')
+  f.close
+
   make_macros(macros_loc_debug, "")  # leave debug(x) as is in debug build
   # replace debug(x) with nothing in release build
   make_macros(macros_loc_default, "macro debug(x) = ;\n")
+  make_macros(macros_loc_default, "macro assert(x) = ;\n")
+
+  if not bld.env["USE_DTRACE"]:
+    probes = [
+      'DTRACE_HTTP_CLIENT_REQUEST',
+      'DTRACE_HTTP_CLIENT_RESPONSE',
+      'DTRACE_HTTP_SERVER_REQUEST',
+      'DTRACE_HTTP_SERVER_RESPONSE',
+      'DTRACE_NET_SERVER_CONNECTION',
+      'DTRACE_NET_STREAM_END',
+      'DTRACE_NET_SOCKET_READ',
+      'DTRACE_NET_SOCKET_WRITE'
+    ]
+
+    for probe in probes:
+      make_macros(macros_loc_default, "macro %s(x) = ;\n" % probe)
+      make_macros(macros_loc_debug, "macro %s(x) = ;\n" % probe)
 
   def javascript_in_c(task):
     env = task.env
@@ -476,41 +649,107 @@ def build(bld):
     native_cc_debug.rule = javascript_in_c_debug
 
   native_cc.rule = javascript_in_c
+  
+  if bld.env["USE_DTRACE"]:
+    dtrace = bld.new_task_gen(
+      name   = "dtrace",
+      source = "src/node_provider.d",
+      target = "src/node_provider.h",
+      rule   = "%s -x nolibs -h -o ${TGT} -s ${SRC}" % (bld.env.DTRACE),
+      before = "cxx",
+    )
+
+    if bld.env["USE_DEBUG"]:
+      dtrace_g = dtrace.clone("debug")
+
+    bld.install_files('${PREFIX}/usr/lib/dtrace', 'src/node.d')
+
+    if sys.platform.startswith("sunos"):
+      #
+      # The USDT DTrace provider works slightly differently on Solaris than on
+      # the Mac; on Solaris, any objects that have USDT DTrace probes must be
+      # post-processed with the DTrace command.  (This is not true on the
+      # Mac, which has first-class linker support for USDT probes.)  On
+      # Solaris, we must therefore post-process our object files.  Waf doesn't
+      # seem to really have a notion for this, so we inject a task after
+      # compiling and before linking, and then find all of the node object
+      # files and shuck them off to dtrace (which will modify them in place
+      # as appropriate).
+      #
+      def dtrace_postprocess(task):
+        abspath = bld.srcnode.abspath(bld.env_of_name(task.env.variant()))
+        objs = glob.glob(abspath + 'src/*.o')
+
+        Utils.exec_command('%s -G -x nolibs -s %s %s' % (task.env.DTRACE,
+          task.inputs[0].srcpath(task.env), ' '.join(objs)))
+
+      dtracepost = bld.new_task_gen(
+        name   = "dtrace-postprocess",
+        source = "src/node_provider.d",
+        always = True,
+        before = "cxx_link",
+        after  = "cxx",
+      )
+
+      bld.env.append_value('LINKFLAGS', 'node_provider.o')
+
+      #
+      # Note that for the same (mysterious) issue outlined above with respect
+      # to assigning the rule to native_cc/native_cc_debug, we must apply the
+      # rule to dtracepost/dtracepost_g only after they have been cloned.  We
+      # also must put node_provider.o on the link line, but because we
+      # (apparently?) lack LINKFLAGS in debug, we (shamelessly) stowaway on
+      # LINKFLAGS_V8_G.
+      #
+      if bld.env["USE_DEBUG"]:
+        dtracepost_g = dtracepost.clone("debug")
+        dtracepost_g.rule = dtrace_postprocess
+        bld.env_of_name("debug").append_value('LINKFLAGS_V8_G',
+          'node_provider.o') 
+
+      dtracepost.rule = dtrace_postprocess
 
   ### node lib
-  node = bld.new_task_gen("cxx", "program")
+  node = bld.new_task_gen("cxx", product_type)
   node.name         = "node"
   node.target       = "node"
-  node.uselib = 'RT EV OPENSSL CARES EXECINFO DL KVM SOCKET NSL'
+  node.uselib = 'RT EV OPENSSL CARES EXECINFO DL KVM SOCKET NSL UTIL OPROFILE'
   node.add_objects = 'eio http_parser'
-  node.install_path = '${PREFIX}/lib'
-  node.install_path = '${PREFIX}/bin'
+  if product_type_is_lib:
+    node.install_path = '${PREFIX}/lib'
+  else:
+    node.install_path = '${PREFIX}/bin'
   node.chmod = 0755
   node.source = """
     src/node.cc
     src/node_buffer.cc
+    src/node_javascript.cc
     src/node_extensions.cc
     src/node_http_parser.cc
     src/node_net.cc
     src/node_io_watcher.cc
-    src/node_child_process.cc
     src/node_constants.cc
     src/node_cares.cc
     src/node_events.cc
     src/node_file.cc
     src/node_signal_watcher.cc
     src/node_stat_watcher.cc
-    src/node_stdio.cc
     src/node_timer.cc
     src/node_script.cc
+    src/node_os.cc
+    src/node_dtrace.cc
   """
 
-  platform_file = "src/platform_%s.cc" % bld.env['DEST_OS']
-  if os.path.exists(join(cwd, platform_file)):
-    node.source += platform_file
+  if sys.platform.startswith("win32"):
+    node.source += " src/node_stdio_win32.cc "
+    node.source += " src/node_child_process_win32.cc "
   else:
-    node.source += "src/platform_none.cc "
+    node.source += " src/node_stdio.cc "
+    node.source += " src/node_child_process.cc "
 
+  node.source += bld.env["PLATFORM_FILE"]
+  if not product_type_is_lib:
+    node.source = 'src/node_main.cc '+node.source
 
   if bld.env["USE_OPENSSL"]: node.source += " src/node_crypto.cc "
 
@@ -540,7 +779,8 @@ def build(bld):
     x = { 'CCFLAGS'   : " ".join(program.env["CCFLAGS"]).replace('"', '\\"')
         , 'CPPFLAGS'  : " ".join(program.env["CPPFLAGS"]).replace('"', '\\"')
         , 'LIBFLAGS'  : " ".join(program.env["LIBFLAGS"]).replace('"', '\\"')
-        , 'PREFIX'    : program.env["PREFIX"]
+        , 'PREFIX'    : safe_path(program.env["PREFIX"])
+        , 'VERSION'   : '0.4.0' # FIXME should not be hard-coded, see NODE_VERSION_STRING in src/node_version.
         }
     return x
 
@@ -578,9 +818,17 @@ def build(bld):
   if os.path.exists('doc/node.1'):
     bld.install_files('${PREFIX}/share/man/man1/', 'doc/node.1')
 
-  bld.install_files('${PREFIX}/bin/', 'bin/*', chmod=0755)
+  bld.install_files('${PREFIX}/bin/', 'tools/node-waf', chmod=0755)
   bld.install_files('${PREFIX}/lib/node/wafadmin', 'tools/wafadmin/*.py')
   bld.install_files('${PREFIX}/lib/node/wafadmin/Tools', 'tools/wafadmin/Tools/*.py')
+
+  # create a pkg-config(1) file
+  node_conf = bld.new_task_gen('subst', before="cxx")
+  node_conf.source = 'tools/nodejs.pc.in'
+  node_conf.target = 'tools/nodejs.pc'
+  node_conf.dict = subflags(node)
+
+  bld.install_files('${PREFIX}/lib/pkgconfig', 'tools/nodejs.pc')
 
 def shutdown():
   Options.options.debug
@@ -590,11 +838,25 @@ def shutdown():
     if not Options.options.use_openssl:
       print "WARNING WARNING WARNING"
       print "OpenSSL not found. Will compile Node without crypto support!"
+
+    if not Options.options.platform_file:
+      print "WARNING: Platform not fully supported. Using src/platform_none.cc"
+
   elif not Options.commands['clean']:
-    if os.path.exists('build/default/node') and not os.path.exists('node'):
-      os.symlink('build/default/node', 'node')
-    if os.path.exists('build/debug/node_g') and not os.path.exists('node_g'):
-      os.symlink('build/debug/node_g', 'node_g')
+    if sys.platform.startswith("win32"):
+      if os.path.exists('build/default/node.exe'):
+        os.system('cp build/default/node.exe .')
+      if os.path.exists('build/debug/node_g.exe'):
+        os.system('cp build/debug/node_g.exe .')
+    else:
+      if os.path.exists('build/default/node') and not os.path.exists('node'):
+        os.symlink('build/default/node', 'node')
+      if os.path.exists('build/debug/node_g') and not os.path.exists('node_g'):
+        os.symlink('build/debug/node_g', 'node_g')
   else:
-    if os.path.exists('node'): os.unlink('node')
-    if os.path.exists('node_g'): os.unlink('node_g')
+    if sys.platform.startswith("win32"):
+      if os.path.exists('node.exe'): os.unlink('node.exe')
+      if os.path.exists('node_g.exe'): os.unlink('node_g.exe')
+    else:
+      if os.path.exists('node'): os.unlink('node')
+      if os.path.exists('node_g'): os.unlink('node_g')
